backend/requirements.txt
fastapi
uvicorn
sqlalchemy
beautifulsoup4
requests
google-generativeai
python-dotenv
pydantic
groq
psycopg2-binary
